pyproject.toml
[tool.poetry]
name = "diquark"
version = "0.1.0"
description = ""
authors = ["Ioan-Mihail Dinu <ioan-mihail.dinu@cern.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9,<3.12"
pandas = "^2.1.3"
awkward = "^2.4.10"
tqdm = "^4.66.1"
uproot = "^5.1.2"
plotly = "^5.18.0"
scikit-learn = "^1.3.2"
joblib = "^1.3.2"
jupyter = "^1.0.0"
scipy = "^1.11.3"
fastparquet = "^2023.10.1"
pendulum = "^2.1.2"
tensorflow = "^2.14.0"
kaleido = "0.2.1"
xgboost = "^2.0.3"
matplotlib = "^3.9.1"
pyyaml = "^6.0.2"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"

[tool.black]
line_length = 100

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
